Re-written librom crt0.s ROM header moved into include file Separated _sbrk into gba-sbrk.c Cleaned up gba-nosyscalls.c atexit and lock data and bss uses EWRAM
