+++ lib/gba-nano.specs
