+++ lib/rom/crt0.s
@@ -9,28 +9,18 @@
 ===============================================================================
 */
 
-    .section .crt0, "ax"
     .align 2
-    .arm
-    .global _start
-_start:
-    // Jump past header to start of ROM
-    b       .Lrom_start
+    .section .crt0, "ax"
 
-    // GBA header
-    .fill   156, 1, 0 // Logo data
-    .fill   12, 1, 0 // ROM title
-    .byte   0x00, 0x00, 0x00, 0x00 // ROM code in UTTD format (U = unique code, TT = short title, D = destination/language)
-    .byte   0x00, 0x00 // Developer ID
-    .byte   0x96 // Fixed value
-    .byte   0x00 // Target device code (0 for GBA models)
-    .byte   0x00 // Debug flags
-    .byte   0x00, 0x00, 0x00 // Unused 3 bytes
-.Lzero_word:
-    .word   0x00000000 // Unused 4 bytes (runtime assumes zero filled)
-    .byte   0x00 // Software version
-    .byte   0x00 // Compliment check
-    .short  0x0000 // Reserved area (should be zero filled)
+    // ROM entry point
+    .arm
+    .global __start
+__start:
+    // Jump past header to start of ROM
+    b       _start
+
+    // Include header data (192 bytes)
+.include "../gba-header.s"
 
     // __ROM_START_PADDING__ bytes can be inserted before rom_start
     // This may be used for save strings
@@ -39,96 +29,129 @@ _start:
     .align  2
 #endif
 
-    // ROM (ARM) execution entry point
+    // _start
     .arm
-.Lrom_start:
-    // r3 set to REG_BASE
-    mov     r3, #0x4000000
-
-    // Disable IME by setting lowest bit to zero (using REG_BASE)
-    str     r3, [r3, #0x208]
-
-    // Switch to IRQ mode (0x12)
-    mov     r0, #0x12
-    msr     cpsr, r0
-    ldr     sp, =__sp_irq // Set IRQ stack pointer
-
-    // Switch to user mode (0x1f)
-    mov     r0, #0x1f
-    msr     cpsr, r0
-    ldr     sp, =__sp_usr // Set user stack pointer
-
+    .global _start
+_start:
     // Enter thumb mode (bit 0 is set to 1)
     adr     r0, .Lthumb_start + 1
     bx      r0
 
-    // ROM (thumb) execution entry point
-    .thumb
+    .thumb_func
 .Lthumb_start:
-    // CpuSet fill bss
+    // CpuSet clear bss (iwram)
     ldr     r0, =.Lzero_word
     ldr     r1, =__bss_start
     ldr     r2, =__bss_cpuset_fill
     swi     #0xb
 
-    // CpuSet fill sbss
+    // CpuSet clear sbss (ewram)
+    ldr     r0, =.Lzero_word
     ldr     r1, =__sbss_start
     ldr     r2, =__sbss_cpuset_fill
     swi     #0xb
 
-    // CpuSet copy ewram
+    // CpuSet copy ewram sections
     ldr     r0, =__ewram_cpuset
     ldm     r0, {r0-r2}
     swi     #0xb
 
-    // CpuSet copy iwram
+    // CpuSet copy iwram sections
     ldr     r0, =__iwram_cpuset
     ldm     r0, {r0-r2}
     swi     #0xb
 
-    // Initializers
-    .extern __libc_init_array
-    ldr     r2, =__libc_init_array
-    bl      .Lbx_r2
+    // preinit_array (r0-r3 are clobbered, r4 is used for .Lbx_r4, so use r5-r6)
+    ldr     r5, =__preinit_array_start
+    ldr     r6, =__preinit_array_end
+    bl      .Lcall_array
 
-    // Main
+    // init_array (r0-r3 are clobbered, r4 is used for .Lbx_r4, so use r5-r6)
+    ldr     r5, =__init_array_start
+    ldr     r6, =__init_array_end
+    bl      .Lcall_array
+
+    // main
     mov     r0, #0 // argc
-    mov     r1, #0 // argv
-    .extern main
-    ldr     r2, =main
-    bl      .Lbx_r2
+    mov     r1, #0 // argv (NULL)
+    ldr     r4, =main
+    bl      .Lbx_r4
+    push    {r0, r1} // Push exit code (r1 for alignment)
 
-    // Finalizers
-    .extern __libc_fini_array
-    ldr     r2, =__libc_fini_array
-    push    {r0}
-    bl      .Lbx_r2
-    pop     {r0}
+    // Disable REG_IME by setting lowest bit to zero (using lowest bit of REG_IME)
+    ldr     r1, =#0x4000208
+    str     r1, [r1]
 
+    // fini_array
+    ldr     r5, =__fini_array_start
+    ldr     r6, =__fini_array_end
+    bl      .Lrcall_array
+
+    pop     {r0, r1} // Pop exit code (r1 for alignment)
+    b       exit
+
+    .thumb_func
+.Lpop_r3_bx_r4:
+    pop     {r3-r4}
+    // Fallthrough to .Lbx_r4
+
+    .thumb_func
+.Lbx_r4:
+    bx      r4
+
+    // call_array (r4 is used for .Lpop_r3_bx_r4, r5-r6 is used for array)
+    .thumb_func
+.Lcall_array:
+    push    {r3, lr} // r3 pushed for alignment
+.Lcall_array_loop:
+    cmp     r5, r6
+    beq     .Lpop_r3_bx_r4
+    ldr     r4, [r5]
+    add     r5, #4
+    bl      .Lbx_r4
+    b       .Lcall_array_loop
+
+    // rcall_array (r4 is used for .Lpop_r3_bx_r4, r5-r6 is used for array)
+    .thumb_func
+.Lrcall_array:
+    push    {r3, lr} // r3 pushed for alignment
+.Lrcall_array_loop:
+    cmp     r5, r6
+    beq     .Lpop_r3_bx_r4
+    sub     r6, #4
+    ldr     r4, [r6]
+    bl      .Lbx_r4
+    b       .Lrcall_array_loop
+
+    // exit
+    .thumb_func
+    .global exit
+    .func exit
+exit:
+    push    {r0-r1} // Push exit code (r1 for alignment)
+
+    // Disable REG_IME by setting lowest bit to zero (using lowest bit of REG_IME)
+    ldr     r1, =#0x4000208
+    str     r1, [r1]
+
+    mov     r1, #0 // NULL
+    ldr     r4, =__call_exitprocs
+    bl      .Lbx_r4
+    pop     {r0, r1} // Pop exit code (r1 for alignment)
     // Fallthrough to _exit
+
+    // _exit
     .thumb_func
     .global _exit
 _exit:
+    // Disable REG_IME by setting lowest bit to zero (using lowest bit of REG_IME)
+    ldr     r0, =#0x4000208
+    str     r0, [r0]
+
     swi     #0x00 // Soft reset
 
-    // For branch-link
-    .thumb
-.Lbx_r2:
-    bx      r2
-
-    // Reference symbols from gba-syscalls.c to prevent removal
+    // Reference _sbrk, _getpid to prevent removal
     .global _sbrk
-    .global _kill
-
-    .section .crt0._getpid, "ax", %progbits
-    .thumb_func
     .global _getpid
-_getpid:
-    mov     r0, #0
-    // Fallthrough to __sync_synchronize
-
-    .section .crt0.__sync_synchronize, "ax", %progbits
-    .thumb_func
-    .global __sync_synchronize
-__sync_synchronize:
-    bx      lr
+    // Reference __register_exitproc to enable static destructors atexit
+    .global __register_exitproc
